+++ src/app/dashboad/products/products.component.ts
@@ -1,12 +1,29 @@
+import { CommonModule } from '@angular/common';
+import { HttpClientModule } from '@angular/common/http';
 import { Component } from '@angular/core';
+import { FormsModule } from '@angular/forms';
+import { ProductsService } from '../../services/products.service';
+import { Products } from '../../models/products';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-products',
   standalone: true,
-  imports: [],
+  imports: [FormsModule, CommonModule, HttpClientModule],
   templateUrl: './products.component.html',
   styleUrl: './products.component.css'
 })
 export class ProductsComponent {
+
+
+  products: Products[] = [];
+
+  constructor(private productsService: ProductsService, private router: Router) { }
+
+  ngOnInit(): void {
+    this.productsService.getAllProducts().subscribe(data => {
+      this.products = data;
+    });
+  }
 
 }
