+++ src/app/models/products.ts
@@ -1,0 +1,10 @@
+export interface Products{
+    id_product: number;
+    name: string;
+    reference: string;
+    price: number;
+    stock: number;
+    brand: string;
+    size: string;
+    image_url: string;
+  }
