+++ src/app/services/products.service.ts
@@ -1,31 +1,24 @@
-import { HttpClient } from '@angular/common/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
-
-
-export interface Products{
-  name: string;
-  reference: string;
-  price: number;
-  stock: number;
-  brand: string;
-  size: string;
-  image: string;
-}
+import { catchError, Observable, tap, throwError } from 'rxjs';
+import { environment } from '../../environments/environment';
+import { Products } from '../models/products';
 
 
 @Injectable({
   providedIn: 'root'
 })
 export class ProductsService {
-  private apiUrl = "http://localhost:8080/products";
+
+  private baseUrl = environment.apiBaseUrl;
 
   constructor(private http: HttpClient) { }
 
   // Método para obtener todos: User
   getAllProducts(): Observable<Products[]> {
-    return this.http.get<Products[]>(this.apiUrl)
-}
+    const headers = new HttpHeaders({ 'Content-Type': 'application/json' });
+    return this.http.get<Products[]>(`${this.baseUrl}/products`, { headers });
+  }
 
 
 }
