+++ src/environments/environment.development.ts
@@ -1,4 +1,4 @@
 export const environment = {
     production: true,
-    apiBaseUrl: 'https://e372-2800-e2-7080-1d1e-24ce-7bc5-998a-de12.ngrok-free.app',
+    apiBaseUrl: 'https://8e90-2800-e2-7080-1d1e-24ce-7bc5-998a-de12.ngrok-free.app',
 };
